feat:Adds "Made in Australia" page and updates menu
Adds a new "Made in Australia" page detailing the company's commitment to Australian manufacturing and its alignment with national priorities.

Updates the navigation menu across multiple pages to include a link to the new "Made in Australia" page, ensuring easy access for users.

Updates the index page with two buttons with responsiveness
Replaces the images in the made in australia pages with new ones, including descriptions